Better tick handle (and worse); q for quit
So now I have it so that there are 20 ticks per second. So everything
can be based on that metric. But... what if I want to change the number
of ticks per second? Should I? Maybe I should set it to 50 or something
instead? But what if the Arduino can't push that fast?

Problems... problems...